Added tester files. Moved assembly tests to its own file. Adding unit tests for assembly instructions.
